pull certain functions into helpers; update api.py to be semi-functional
